fix: increase battery cutout
I increased the size of the battery cutout to
ensure the battery clips fit in the cutout. I left
the battery clips where they were, since they are
42.99mm apart, which is 1mm less than what the
datasheet recommends for a AAA battery. However,
the 10440 battery I will be using is 1mm smaller
than a AAA, so this should be fine.