fix: flip ground solder jumper pad
The MCU's GND solder jumper pad was incorrectly
flipped in the PCB layout, requiring it to be
bridged on the wrong side.

This commit corrects the orientation of the solder
jumper, so that it can be bridged on the correct
side. Now, solder jumpers should all be bridged on
the back of the PCB, regardless of which side of
the keyboard the PCB is used for.